Fixed
Seriously , int on checksum and File table? Now DB Table ticket is complete. God bless and VR tommorow :)

+++ database/tables/files.sql
@@ -3,6 +3,6 @@ CREATE TABLE files (
     path TEXT NOT NULL UNIQUE,
     basename TEXT NOT NULL,
     directory TEXT NOT NULL,
-    size BIGINT NOT NULL,
+    size INTEGER NOT NULL,
     mtime TIMESTAMPTZ NOT NULL
 );
