+++ database/tables/scan_details.sql
@@ -1,6 +1,6 @@
 CREATE TABLE scan_details (
-    id BIGSERIAL PRIMARY KEY,
-    scan_id bigint NOT NULL REFERENCES scans(id),
+    id SERIAL PRIMARY KEY,
+    scan_id INTEGER NOT NULL REFERENCES scans(id),
     file_id bigint NOT NULL REFERENCES files(id),
     checksum_id bigint NOT NULL REFERENCES checksums(id)
 );
