+++ database/tables/scans.sql
@@ -1,7 +1,7 @@
 CREATE TABLE scans (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     scan_time TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     status TEXT NOT NULL,
     notes TEXT,
-    monitored_directory_id bigint NOT NULL REFERENCES monitored_directories(id)
+    monitored_directory_id INTEGER NOT NULL REFERENCES monitored_directories(id)
 );
